Create a dedicated crate
