+++ crates/uv-build-frontend/src/lib.rs
@@ -32,8 +32,8 @@ use uv_configuration::PreviewMode;
 use uv_configuration::{BuildKind, BuildOutput, ConfigSettings, SourceStrategy};
 use uv_distribution::BuildRequires;
 use uv_distribution_types::{IndexLocations, Requirement, Resolution};
-use uv_fs::LockedFile;
 use uv_fs::{PythonExt, Simplified};
+use uv_lock::LockedFile;
 use uv_pep440::Version;
 use uv_pep508::PackageName;
 use uv_pypi_types::VerbatimParsedUrl;
