+++ crates/uv-fs/src/lib.rs
@@ -1,10 +1,8 @@
-use std::fmt::Display;
 use std::io;
 use std::path::{Path, PathBuf};
 
-use fs2::FileExt;
 use tempfile::NamedTempFile;
-use tracing::{debug, error, info, trace, warn};
+use tracing::warn;
 
 pub use crate::path::*;
 
@@ -597,137 +595,6 @@ pub fn is_virtualenv_executable(executable: impl AsRef<Path>) -> bool {
 /// unnecessary.
 pub fn is_virtualenv_base(path: impl AsRef<Path>) -> bool {
     path.as_ref().join("pyvenv.cfg").is_file()
-}
-
-/// A file lock that is automatically released when dropped.
-#[derive(Debug)]
-#[must_use]
-pub struct LockedFile(fs_err::File);
-
-impl LockedFile {
-    /// Inner implementation for [`LockedFile::acquire_blocking`] and [`LockedFile::acquire`].
-    fn lock_file_blocking(file: fs_err::File, resource: &str) -> Result<Self, std::io::Error> {
-        trace!(
-            "Checking lock for `{resource}` at `{}`",
-            file.path().user_display()
-        );
-        match file.file().try_lock_exclusive() {
-            Ok(()) => {
-                debug!("Acquired lock for `{resource}`");
-                Ok(Self(file))
-            }
-            Err(err) => {
-                // Log error code and enum kind to help debugging more exotic failures.
-                if err.kind() != std::io::ErrorKind::WouldBlock {
-                    debug!("Try lock error: {err:?}");
-                }
-                info!(
-                    "Waiting to acquire lock for `{resource}` at `{}`",
-                    file.path().user_display(),
-                );
-                file.file().lock_exclusive().map_err(|err| {
-                    // Not an fs_err method, we need to build our own path context
-                    std::io::Error::other(format!(
-                        "Could not acquire lock for `{resource}` at `{}`: {}",
-                        file.path().user_display(),
-                        err
-                    ))
-                })?;
-
-                debug!("Acquired lock for `{resource}`");
-                Ok(Self(file))
-            }
-        }
-    }
-
-    /// The same as [`LockedFile::acquire`], but for synchronous contexts. Do not use from an async
-    /// context, as this can block the runtime while waiting for another process to release the
-    /// lock.
-    pub fn acquire_blocking(
-        path: impl AsRef<Path>,
-        resource: impl Display,
-    ) -> Result<Self, std::io::Error> {
-        let file = Self::create(path)?;
-        let resource = resource.to_string();
-        Self::lock_file_blocking(file, &resource)
-    }
-
-    /// Acquire a cross-process lock for a resource using a file at the provided path.
-    #[cfg(feature = "tokio")]
-    pub async fn acquire(
-        path: impl AsRef<Path>,
-        resource: impl Display,
-    ) -> Result<Self, std::io::Error> {
-        let file = Self::create(path)?;
-        let resource = resource.to_string();
-        tokio::task::spawn_blocking(move || Self::lock_file_blocking(file, &resource)).await?
-    }
-
-    #[cfg(unix)]
-    fn create(path: impl AsRef<Path>) -> Result<fs_err::File, std::io::Error> {
-        use std::os::unix::fs::PermissionsExt;
-
-        // If path already exists, return it.
-        if let Ok(file) = fs_err::OpenOptions::new()
-            .read(true)
-            .write(true)
-            .open(path.as_ref())
-        {
-            return Ok(file);
-        }
-
-        // Otherwise, create a temporary file with 777 permissions. We must set
-        // permissions _after_ creating the file, to override the `umask`.
-        let file = if let Some(parent) = path.as_ref().parent() {
-            NamedTempFile::new_in(parent)?
-        } else {
-            NamedTempFile::new()?
-        };
-        if let Err(err) = file
-            .as_file()
-            .set_permissions(std::fs::Permissions::from_mode(0o777))
-        {
-            warn!("Failed to set permissions on temporary file: {err}");
-        }
-
-        // Try to move the file to path, but if path exists now, just open path
-        match file.persist_noclobber(path.as_ref()) {
-            Ok(file) => Ok(fs_err::File::from_parts(file, path.as_ref())),
-            Err(err) => {
-                if err.error.kind() == std::io::ErrorKind::AlreadyExists {
-                    fs_err::OpenOptions::new()
-                        .read(true)
-                        .write(true)
-                        .open(path.as_ref())
-                } else {
-                    Err(err.error)
-                }
-            }
-        }
-    }
-
-    #[cfg(not(unix))]
-    fn create(path: impl AsRef<Path>) -> std::io::Result<fs_err::File> {
-        fs_err::OpenOptions::new()
-            .read(true)
-            .write(true)
-            .create(true)
-            .open(path.as_ref())
-    }
-}
-
-impl Drop for LockedFile {
-    fn drop(&mut self) {
-        if let Err(err) = fs2::FileExt::unlock(self.0.file()) {
-            error!(
-                "Failed to unlock {}; program may be stuck: {}",
-                self.0.path().display(),
-                err
-            );
-        } else {
-            debug!("Released lock at `{}`", self.0.path().display());
-        }
-    }
 }
 
 /// An asynchronous reader that reports progress as bytes are read.
