+++ crates/uv-tool/Cargo.toml
@@ -23,6 +23,7 @@ uv-distribution-types = { workspace = true }
 uv-fs = { workspace = true }
 uv-install-wheel = { workspace = true }
 uv-installer = { workspace = true }
+uv-lock = { workspace = true, features = ["tokio"] }
 uv-pep440 = { workspace = true }
 uv-pep508 = { workspace = true }
 uv-pypi-types = { workspace = true }
