+++ crates/uv-tool/src/lib.rs
@@ -19,8 +19,9 @@ use uv_install_wheel::read_record_file;
 pub use receipt::ToolReceipt;
 pub use tool::{Tool, ToolEntrypoint};
 use uv_cache::Cache;
-use uv_fs::{LockedFile, Simplified};
+use uv_fs::Simplified;
 use uv_installer::SitePackages;
+use uv_lock::LockedFile;
 use uv_python::{Interpreter, PythonEnvironment};
 use uv_state::{StateBucket, StateStore};
 use uv_static::EnvVars;
